server/requirements.txt
# Core Framework
fastapi==0.109.1
uvicorn[standard]==0.27.1
gunicorn==21.2.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# Data Processing & ML
pandas==2.2.3
numpy==1.26.4
prophet==1.1.5
scikit-learn==1.6.1
joblib==1.3.2

# Data Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0

# Development Tools
black==24.1.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
aiohttp==3.9.3
APScheduler==3.10.4

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.9
